Allow usage of fs within the app, disable zmq.
ZMQ uses bindings which is incompatible with webpack.
Details: bindings tries to reach the directory structure found in the development directory, which is not the case in the dist directory.
More details: bindings uses the prepareStackTrace to examine the trace files, and webpack calls zmq--(thus)-->bindings.js through eval which doesn't result in having the filenames accessible through getEvalOrigin() instead of getFileName() follow [this issue](https://github.com/TooTallNate/node-bindings/issues/50).